Split osm store into separate file
