+++ src/osm_reader/osm_utils.rs
@@ -31,7 +31,7 @@
 use mimir;
 use osmpbfreader;
 
-use super::street::Getter;
+use super::osm_store::Getter;
 use geo::centroid::Centroid;
 use geo::MultiPolygon;
 use osmpbfreader::StoreObjs;
