+++ src/osm_reader/street.rs
@@ -27,345 +27,49 @@
 // IRC #navitia on freenode
 // https://groups.google.com/d/forum/navitia
 // www.navitia.io
+
+use super::osm_store::{Getter, Kind, ObjWrapper};
 use super::osm_utils::get_way_coord;
 use super::OsmPbfReader;
 use crate::admin_geofinder::AdminGeoFinder;
 use crate::{labels, utils, Error};
-use bincode;
 use failure::ResultExt;
-use osmpbfreader::{OsmId, OsmObj, StoreObjs};
-use rusqlite::{Connection, DropBehavior, ToSql, NO_PARAMS};
-use slog_scope::{error, info};
-use std::borrow::Cow;
-use std::collections::{BTreeMap, BTreeSet, HashMap};
-use std::fs;
+use osmpbfreader::StoreObjs;
+use slog_scope::{debug, info};
+use std::collections::{BTreeMap, BTreeSet};
 use std::ops::Deref;
 use std::path::PathBuf;
 use std::sync::Arc;
 
-macro_rules! err_logger {
-    ($obj:expr, $err_msg:expr) => {
-        match $obj {
-            Ok(x) => Some(x),
-            Err(e) => {
-                error!("{}: {}", $err_msg, e);
-                None
-            }
-        }?
-    };
-    ($obj:expr, $err_msg:expr, $ret:expr) => {
-        match $obj {
-            Ok(x) => x,
-            Err(e) => {
-                error!("{}: {}", $err_msg, e);
-                return $ret;
-            }
-        }
-    };
-}
-
-macro_rules! get_kind {
-    ($obj:expr) => {
-        if $obj.is_node() {
-            &0
-        } else if $obj.is_way() {
-            &1
-        } else if $obj.is_relation() {
-            &2
-        } else {
-            panic!("Unknown OSM object kind!")
-        }
-    };
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
-#[allow(dead_code)]
-enum Kind {
-    Node = 0,
-    Way = 1,
-    Relation = 2,
-}
-
-pub trait Getter {
-    fn get(&self, key: &OsmId) -> Option<Cow<OsmObj>>;
-}
-
-impl Getter for BTreeMap<OsmId, OsmObj> {
-    fn get(&self, key: &OsmId) -> Option<Cow<OsmObj>> {
-        self.get(key).map(|x| Cow::Borrowed(x))
-    }
-}
-
-struct DB<'a> {
-    conn: Connection,
-    db_file: &'a PathBuf,
-    buffer: HashMap<OsmId, OsmObj>,
-    db_buffer_size: usize,
-}
-
-impl<'a> DB<'a> {
-    fn new(db_file: &'a PathBuf, db_buffer_size: usize) -> Result<DB<'a>, String> {
-        let _ = fs::remove_file(db_file); // we ignore any potential error
-        let conn = Connection::open(&db_file)
-            .map_err(|e| format!("failed to open SQLITE connection: {}", e))?;
-
-        conn.execute(
-            "CREATE TABLE ids (
-                id   INTEGER NOT NULL,
-                obj  BLOB NOT NULL,
-                kind INTEGER NOT NULL,
-                UNIQUE(id, kind)
-             )",
-            NO_PARAMS,
-        )
-        .map_err(|e| format!("failed to create table: {}", e))?;
-        Ok(DB {
-            conn,
-            db_file,
-            buffer: HashMap::with_capacity(db_buffer_size),
-            db_buffer_size,
-        })
-    }
-
-    fn get_from_id(&self, id: &OsmId) -> Option<Cow<OsmObj>> {
-        if let Some(obj) = self.buffer.get(id) {
-            return Some(Cow::Borrowed(obj));
-        }
-        let mut stmt = err_logger!(
-            self.conn
-                .prepare("SELECT obj FROM ids WHERE id=?1 AND kind=?2"),
-            "DB::get_from_id: prepare failed"
-        );
-        let mut iter = err_logger!(
-            stmt.query(&[&id.inner_id() as &dyn ToSql, get_kind!(id)]),
-            "DB::get_from_id: query_map failed"
-        );
-        while let Some(row) = err_logger!(iter.next(), "DB::get_from_id: next failed") {
-            let obj: Vec<u8> = err_logger!(row.get(0), "DB::get_from_id: failed to get obj field");
-            let obj: OsmObj = err_logger!(
-                bincode::deserialize(&obj),
-                "DB::for_each: serde conversion failed",
-                None
-            );
-            return Some(Cow::Owned(obj));
-        }
-        None
-    }
-
-    #[allow(dead_code)]
-    fn for_each<F: FnMut(Cow<OsmObj>)>(&self, mut f: F) {
-        let mut values = self.buffer.values();
-        while let Some(obj) = values.next() {
-            f(Cow::Borrowed(obj));
-        }
-        let mut stmt = err_logger!(
-            self.conn.prepare("SELECT obj FROM ids"),
-            "DB::for_each: prepare failed",
-            ()
-        );
-        let mut rows = err_logger!(stmt.query(NO_PARAMS), "DB::for_each: query_map failed", ());
-        while let Some(row) = err_logger!(rows.next(), "DB::for_each: next failed", ()) {
-            let obj: Vec<u8> = err_logger!(row.get(0), "DB::for_each: failed to get obj field", ());
-
-            let obj: OsmObj = err_logger!(
-                bincode::deserialize(&obj),
-                "DB::for_each: serde conversion failed",
-                ()
-            );
-            f(Cow::Owned(obj));
-        }
-    }
-
-    fn for_each_filter<F: FnMut(Cow<OsmObj>)>(&self, filter: Kind, mut f: F) {
-        self.buffer
-            .values()
-            .filter(|e| *get_kind!(e) == filter as i32)
-            .for_each(|value| f(Cow::Borrowed(value)));
-        let mut stmt = err_logger!(
-            self.conn.prepare("SELECT obj FROM ids WHERE kind=?1"),
-            "DB::for_each: prepare failed",
-            ()
-        );
-        let mut rows = err_logger!(
-            stmt.query(&[&(filter as i32) as &dyn ToSql]),
-            "DB::for_each: query_map failed",
-            ()
-        );
-        while let Some(row) = err_logger!(rows.next(), "DB::for_each: next failed", ()) {
-            let obj: Vec<u8> = err_logger!(row.get(0), "DB::for_each: failed to get obj field", ());
-
-            let obj: OsmObj = err_logger!(
-                bincode::deserialize(&obj),
-                "DB::for_each: serde conversion failed",
-                ()
-            );
-            f(Cow::Owned(obj));
-        }
-    }
-
-    fn flush_buffer(&mut self) {
-        if self.buffer.is_empty() {
-            return;
-        }
-        let mut tx = err_logger!(
-            self.conn.transaction(),
-            "DB::flush_buffer: transaction creation failed",
-            ()
-        );
-        tx.set_drop_behavior(DropBehavior::Ignore);
-
-        {
-            let mut stmt = err_logger!(
-                tx.prepare("INSERT OR IGNORE INTO ids(id, obj, kind) VALUES (?1, ?2, ?3)"),
-                "DB::flush_buffer: prepare failed",
-                ()
-            );
-            for (id, obj) in self.buffer.drain() {
-                let ser_obj = match bincode::serialize(&obj) {
-                    Ok(s) => s,
-                    Err(e) => {
-                        error!("DB::flush_buffer: failed to convert to json: {}", e);
-                        continue;
-                    }
-                };
-                let kind = get_kind!(obj);
-                if let Err(e) = stmt.execute(&[&id.inner_id() as &dyn ToSql, &ser_obj, kind]) {
-                    error!("DB::flush_buffer: insert failed: {}", e);
-                }
-            }
-        }
-        err_logger!(
-            tx.commit(),
-            "DB::flush_buffer: transaction commit failed",
-            ()
-        );
-    }
-}
-
-impl<'a> StoreObjs for DB<'a> {
-    fn insert(&mut self, id: OsmId, obj: OsmObj) {
-        if self.buffer.len() >= self.db_buffer_size {
-            self.flush_buffer();
-        }
-        self.buffer.insert(id, obj);
-    }
-
-    fn contains_key(&self, id: &OsmId) -> bool {
-        if self.buffer.contains_key(id) {
-            return true;
-        }
-        let mut stmt = err_logger!(
-            self.conn
-                .prepare("SELECT id FROM ids WHERE id=?1 AND kind=?2"),
-            "DB::contains_key: prepare failed",
-            false
-        );
-        let mut iter = err_logger!(
-            stmt.query(&[&id.inner_id() as &dyn ToSql, get_kind!(id)]),
-            "DB::contains_key: query_map failed",
-            false
-        );
-        err_logger!(iter.next(), "DB::contains_key: no row", false).is_some()
-    }
-}
-
-impl<'a> Getter for DB<'a> {
-    fn get(&self, key: &OsmId) -> Option<Cow<OsmObj>> {
-        self.get_from_id(key)
-    }
-}
-
-impl<'a> Drop for DB<'a> {
-    fn drop(&mut self) {
-        let _ = fs::remove_file(self.db_file); // we ignore any potential error
-    }
-}
-
-pub type AdminSet = BTreeSet<Arc<mimir::Admin>>;
-pub type NameAdminMap = BTreeMap<StreetKey, Vec<osmpbfreader::OsmId>>;
-pub type StreetsVec = Vec<mimir::Street>;
-pub type StreetWithRelationSet = BTreeSet<osmpbfreader::OsmId>;
-
+// This structure is used to prevent having multiple streets with the same name
+// in the same admin city level.
 #[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq)]
 pub struct StreetKey {
     pub name: String,
-    pub admins: AdminSet,
+    pub admins: BTreeSet<Arc<mimir::Admin>>,
 }
 
-enum ObjWrapper<'a> {
-    Map(BTreeMap<osmpbfreader::OsmId, osmpbfreader::OsmObj>),
-    DB(DB<'a>),
-}
-
-impl<'a> ObjWrapper<'a> {
-    fn new(db_file: &'a Option<PathBuf>, db_buffer_size: usize) -> Result<ObjWrapper<'a>, Error> {
-        Ok(if let Some(ref db_file) = db_file {
-            info!("Running with DB storage");
-            ObjWrapper::DB(DB::new(db_file, db_buffer_size).map_err(|e| failure::err_msg(e))?)
-        } else {
-            info!("Running with BTreeMap (RAM) storage");
-            ObjWrapper::Map(BTreeMap::new())
-        })
-    }
-
-    #[allow(dead_code)]
-    fn for_each<F: FnMut(Cow<OsmObj>)>(&self, mut f: F) {
-        match *self {
-            ObjWrapper::Map(ref m) => {
-                let mut values = m.values();
-                while let Some(value) = values.next() {
-                    f(Cow::Borrowed(value));
-                }
-            }
-            ObjWrapper::DB(ref db) => db.for_each(f),
-        }
-    }
-
-    fn for_each_filter<F: FnMut(Cow<OsmObj>)>(&self, filter: Kind, mut f: F) {
-        match *self {
-            ObjWrapper::Map(ref m) => {
-                m.values()
-                    .filter(|e| *get_kind!(e) == filter as i32)
-                    .for_each(|value| f(Cow::Borrowed(value)));
-            }
-            ObjWrapper::DB(ref db) => db.for_each_filter(filter, f),
-        }
-    }
-}
-
-impl<'a> Getter for ObjWrapper<'a> {
-    fn get(&self, key: &OsmId) -> Option<Cow<OsmObj>> {
-        match *self {
-            ObjWrapper::Map(ref m) => m.get(key).map(|x| Cow::Borrowed(x)),
-            ObjWrapper::DB(ref db) => db.get(key),
-        }
-    }
-}
-
-impl<'a> StoreObjs for ObjWrapper<'a> {
-    fn insert(&mut self, id: OsmId, obj: OsmObj) {
-        match *self {
-            ObjWrapper::Map(ref mut m) => {
-                m.insert(id, obj);
-            }
-            ObjWrapper::DB(ref mut db) => db.insert(id, obj),
-        }
-    }
-
-    fn contains_key(&self, id: &OsmId) -> bool {
-        match *self {
-            ObjWrapper::Map(ref m) => m.contains_key(id),
-            ObjWrapper::DB(ref db) => db.contains_key(id),
-        }
-    }
-}
-
+// This function generates the list of streets found by reading from the OSM PBF Reader.
+// It does a lot of work trying to merge parts of streets that may be otherwise split in OSM to avoid
+// returning duplicate answers to the user.
+//
+// Sometimes, streets can be divided into several "way"s that still have the same street name.
+// The reason why a street is divided may be that a part of the street become
+// a bridge/tunnel/etc. In this case, a "relation" tagged with (type = associatedStreet) is used
+// to group all these "way"s. In order not to have duplicates in autocompletion, we should tag
+// the osm ways in the relation not to index them twice.
+// The parameter db_file will be used if we use an in memory map to store the streets while we
+// work on them, or a sqlite database (on file).
 pub fn streets(
     pbf: &mut OsmPbfReader,
     admins_geofinder: &AdminGeoFinder,
     db_file: &Option<PathBuf>,
     db_buffer_size: usize,
-) -> Result<StreetsVec, Error> {
+) -> Result<Vec<mimir::Street>, Error> {
+    // First we read from the pbf to extract all the valid ways, which can be either
+    // - ways
+    // - relations with the tag 'associatedStreet'
+
     // This is the list of highway that we don't want to index
     // See [OSM Key Highway](https://wiki.openstreetmap.org/wiki/Key:highway) for background.
     const INVALID_HIGHWAY: &'static [&'static str] =
@@ -373,7 +77,7 @@ pub fn streets(
 
     // For the object to be a valid street, it needs to be an osm highway of a valid type,
     // or a relation of type associatedStreet.
-    fn is_valid_obj(obj: &osmpbfreader::OsmObj) -> bool {
+    fn is_valid_way(obj: &osmpbfreader::OsmObj) -> bool {
         match *obj {
             osmpbfreader::OsmObj::Way(ref way) => {
                 way.tags.get("highway").map_or(false, |v| {
@@ -387,18 +91,16 @@ pub fn streets(
             _ => false,
         }
     }
-    info!("reading pbf...");
+    debug!("reading pbf...");
     let mut objs_map = ObjWrapper::new(db_file, db_buffer_size)?;
-    pbf.get_objs_and_deps_store(is_valid_obj, &mut objs_map)
+    pbf.get_objs_and_deps_store(is_valid_way, &mut objs_map)
         .context("Error occurred when reading pbf")?;
-    info!("reading pbf done.");
-    let mut street_rel: StreetWithRelationSet = BTreeSet::new();
-    let mut street_list: StreetsVec = vec![];
-    // Sometimes, streets can be divided into several "way"s that still have the same street name.
-    // The reason why a street is divided may be that a part of the street become
-    // a bridge/tunnel/etc. In this case, a "relation" tagged with (type = associatedStreet) is used
-    // to group all these "way"s. In order not to have duplicates in autocompletion, we should tag
-    // the osm ways in the relation not to index them twice.
+    debug!("reading pbf done.");
+
+    // Second we build a list of all the streets that are part of a relation.
+    // TODO street_rel
+    let mut street_rel: BTreeSet<osmpbfreader::OsmId> = BTreeSet::new();
+    let mut street_list: Vec<mimir::Street> = vec![];
 
     objs_map.for_each_filter(Kind::Relation, |obj| {
         let rel = obj.relation().expect("impossible unwrap failure occured");
@@ -446,7 +148,7 @@ pub fn streets(
 
     // we merge all the ways with a key = way_name + admin list of level(=city_level)
     // we use a map NameAdminMap <key, value> to manage the merging of ways
-    let mut name_admin_map = NameAdminMap::default();
+    let mut name_admin_map: BTreeMap<StreetKey, Vec<osmpbfreader::OsmId>> = BTreeMap::default();
     objs_map.for_each_filter(Kind::Way, |obj| {
         let osmid = obj.id();
         if street_rel.contains(&osmid) {
@@ -523,7 +225,7 @@ fn get_street_admin<T: StoreObjs + Getter>(
         .map_or(vec![], |c| admins_geofinder.get(&c))
 }
 
-pub fn compute_street_weight(streets: &mut StreetsVec) {
+pub fn compute_street_weight(streets: &mut Vec<mimir::Street>) {
     for st in streets {
         for admin in &mut st.administrative_regions {
             if admin.is_city() {
